Repeating code moved to common.c file, makefile updated.
