+++ Makefile
@@ -11,19 +11,26 @@ CC = gcc
 CFLAGS = -lm -pthread -O3 -march=native -funroll-loops -Wall -Wextra -Wpedantic
 BUILDDIR := build
 SRCDIR := src
+OBJDIR := $(BUILDDIR)
 
-all: dir glove shuffle cooccur vocab_count
+OBJ := $(OBJDIR)/vocab_count.o $(OBJDIR)/cooccur.o $(OBJDIR)/shuffle.o $(OBJDIR)/glove.o
+HEADERS := $(SRCDIR)/common.h
+MODULES := $(BUILDDIR)/vocab_count $(BUILDDIR)/cooccur $(BUILDDIR)/shuffle $(BUILDDIR)/glove
 
+
+all: dir $(OBJ) $(MODULES)
 dir :
 	mkdir -p $(BUILDDIR)
-glove : $(SRCDIR)/glove.c
-	$(CC) $(SRCDIR)/glove.c -o $(BUILDDIR)/glove $(CFLAGS)
-shuffle : $(SRCDIR)/shuffle.c
-	$(CC) $(SRCDIR)/shuffle.c -o $(BUILDDIR)/shuffle $(CFLAGS)
-cooccur : $(SRCDIR)/cooccur.c
-	$(CC) $(SRCDIR)/cooccur.c -o $(BUILDDIR)/cooccur $(CFLAGS)
-vocab_count : $(SRCDIR)/vocab_count.c
-	$(CC) $(SRCDIR)/vocab_count.c -o $(BUILDDIR)/vocab_count $(CFLAGS)
-
+$(BUILDDIR)/glove : $(OBJDIR)/glove.o $(OBJDIR)/common.o
+	$(CC) $^ -o $@ $(CFLAGS)
+$(BUILDDIR)/shuffle : $(OBJDIR)/shuffle.o $(OBJDIR)/common.o
+	$(CC) $^ -o $@ $(CFLAGS)
+$(BUILDDIR)/cooccur : $(OBJDIR)/cooccur.o $(OBJDIR)/common.o
+	$(CC) $^ -o $@ $(CFLAGS)
+$(BUILDDIR)/vocab_count : $(OBJDIR)/vocab_count.o $(OBJDIR)/common.o
+	$(CC) $^ -o $@ $(CFLAGS)
+$(OBJDIR)/%.o : $(SRCDIR)/%.c $(HEADERS)
+	$(CC) -c $< -o $@ $(CFLAGS)
+.PHONY: clean
 clean:
-	rm -rf glove shuffle cooccur vocab_count build
+	rm -rf $(BUILDDIR)
